Unsiged signature and linear test improve
* Unsigned signature for coefficients

Bofore it was possible to add a negative coeffitient with negative
probability meaning so we coult get a negative probability from sum of
coeffitients.

* Default 0 probability for coefficient.

Undefined coeffitient now return 0 instead out of bound error, which is a natural
behavior.

Explaining:
- For example, the first thing probability is 50%.
- 50% for the second.
- Probability of the 132'th item is 0.

* Linear test shows linear result

Before count of generations was too small to let the destribution be monotonous.